COM-1689 - create authentication files
