+++ src/core/api/Users.js
@@ -1,35 +1,7 @@
-import axios from 'axios';
 import { alenviAxios } from '@api/ressources/alenviAxios';
 import { WEBAPP } from '@data/constants';
 
 export default {
-  async refreshToken (data) {
-    await axios.post(`${process.env.API_HOSTNAME}/users/refreshToken`, data, { withCredentials: true });
-  },
-  async authenticate (data) {
-    const auth = await axios.post(`${process.env.API_HOSTNAME}/users/authenticate`, data, { withCredentials: true });
-    return auth.data.data;
-  },
-  async logOut () {
-    await axios.post(`${process.env.API_HOSTNAME}/users/logout`, null, { withCredentials: true });
-  },
-  async forgotPassword (data) {
-    await axios.post(`${process.env.API_HOSTNAME}/users/forgot-password`, data);
-  },
-  async checkResetPasswordToken (resetToken) {
-    const check = await axios.get(
-      `${process.env.API_HOSTNAME}/users/check-reset-password/${resetToken}`,
-      { withCredentials: true }
-    );
-    return check.data.data;
-  },
-  async createPasswordToken (id, data) {
-    const passwordToken = await alenviAxios.put(`${process.env.API_HOSTNAME}/users/${id}/create-password-token`, data);
-    return passwordToken.data.data.passwordToken;
-  },
-  async updatePassword (userId, data) {
-    await alenviAxios.put(`${process.env.API_HOSTNAME}/users/${userId}/password`, data);
-  },
   async list (params = null) {
     try {
       const usersRaw = await alenviAxios.get(`${process.env.API_HOSTNAME}/users`, { params });
